modulo-11/ponderada-semana-3/data_pipeline_abner/pyproject.toml
[tool.poetry]
name = "data-pipeline-abner"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
requests = "^2.32.3"
pandas = "^2.2.2"
pyarrow = "^17.0.0"
minio = "^7.2.8"
clickhouse-connect = "^0.7.19"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
